requirements.txt
numpy>=1.16.5
pandas
scikit-learn
flask
python-dotenv
gunicorn
plotly
statsmodels
matplotlib
typing
pillow
pandas
dash
dash_core_components
dash_bootstrap_components
dash_html_components
tqdm
dash_table
https://download.pytorch.org/whl/cpu/torch-1.7.1%2Bcpu-cp37-cp37m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.8.2%2Bcpu-cp37-cp37m-linux_x86_64.whl